+++ apps/web/src/components/Profile/NftGallery/Picker.tsx
@@ -1,4 +1,4 @@
-import SingleNft from '@components/Nfts/SingleNft';
+import SingleNft from '@components/Nft/SingleNft';
 import NftPickerShimmer from '@components/Shared/Shimmer/NftPickerShimmer';
 import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
